Simple wiki concept framing
When you register you get two bags _<user>-data and <user>-notebook.

Both are "private" (other options include public, protected, custom,
a bit like hoster).

On your dashboard you can see links to all your "tanks" which are
bags that you can do stuff with, initially to be operating as a simple
wiki-in-one-bag.

Later recipes will be used to remix stuff from the various tanks.

There are terms of convenience, subject to change, etc.

It's bedtime. When in future visits:

* editing content
* tiddlers-in-the tank lists
* links between bags/tanks
* drag and drop binaries into tanks
* creating new tanks with misc policies