Add a very simple editor at /edit
GET gets the editor for bag/tiddler args
POST saves the tiddler and redirects (for now) back to the tank page.

Some slight style improvements.

This now gets the login, go to dash, go to private notebook, go
to index, edit index, make pages path complete.

deleting is currently disabled

and there's no functionality for creating other tanks
or for linking between tanks

but this is an mvp, so I'll deploy it tomorrow