Add support for markdown-checklists
Thanks to @FND for an easy to copy example in BFW.

The basic gist here is that if you click on a checkbox the tiddler
is put to the server. Edit conflict handling not yet in place.

For the time being, while using devstore2 there is no need to
worry about asset management, but at some point there will be.