Add simple tag search support using whoosh
Also add a /tags route which list tags used in various contexts.
With no q query parameter it lists all tags. A q=bag:foobar will
list tags in bag foobar. q=bag:foobar modifier:cdent will be just
tags from tiddlers modified by cdent (in bag foobar).

This is not yet robust (the query parsing is week in the face
of quotes and spaces) but is an experiment worth pursuing.

This requires the very latest version of tiddlywebplugins.whoosher
which stores tag strings in the index itself.

/cc @FND for comparison with tagdex (this does something slightly
different).