Add RecentChanges
I've gone with the "special pages" concept. That is, certain page
names short circuit the usual page handling to special functions for
that name.

changes, in this context, is defined as tiddlers modified in the
last n days, where n is 7 by default but other options can be chosen
by a 'd' query parameter.

presentation at this point is intentionally very limited but there's
room for expansion

Doing this work points out there will eventually need to be some
handling of header/menu items.