Introduce the concept of a "comp"
A comp is a composition, which for now is a recipe which contains
at least one tiddler named "app". Loading /comp/foobar looks for
app in the foobar recipe and then displays it (as a single tiddler).

This allows for tanks to be paired with "apps".

This is highly expermental at this point: there is no UI for
creating a comp or recipe.