Include an empty.html of tiddlywiki classic as an "app"
This doesn't work yet, as the tiddlywiki does not know to load
content from the server. However if you have a recipe named foo of

  /bags/tiddlywiki/tiddlers
  /bags/sometank/tiddlers

When going to /comps/foo you get a tiddlywiki.

We need to add TiddlyWebAdaptor and friends, and boostrapping code
to load the initial tiddlers from the right place.

This will likely seem a bit slow, but ought to be no slower than
existing TiddlyWiki handling _and_ it clarifies the slowness as
in TiddlyWiki and give clear room for where improvement needs to
happen.