Establish icons with policy types
fixes #18

This required a fair bit of code moving around and an addition to the
tiddlyweb core to support comparing policies for equality.

The result, though, is nice: We have a simple iconic grammar that can
be used through the system.

In fact an additional change that would be nice is that each tank's
wiki pages ought to display the icon of the policy state.