Add tag autocomplete in editor
fixes #35

This uses selectize, thanks be to @FND for stuff to copy from, like
a good cargo culter.

My use case/implementation is somewhat different from his since
this one is just doing simple autocomplete in the context of either this
bag, or all tags globally.

I had to hack in a global 'tags' var to prevent too many returns to
the server for data. I assume there is a way to prevent this via
options but the docs were not clear...

The data source is at /tags and for the time being uses whoosh.