Allow editing of the bag and its pieces parts.
This is not quite elegant yet, but has a path to being
pretty good. In the future:

* It would be nice if adding members used selectize or had
  similar behavior. There appear to be a variety of ways to
  do it in an angular kind of way, e.g.
      http://jsfiddle.net/sebmade/swfjT/
* It would be nice if it were nice to look at.
* It would be nice if bags were deletable (perhaps)

The primary reason for loading all the bags and their policies at
startup time is to make it possible to have a data set which
is a unique list of all the users being used, so that they can be
used in autocompletion.