Revamp top menu layout and presentation
Mainly this is to fix #91 and show the page name more
clearly on the given page.

However, to make that make sense more of the "command oriented"
links in the top bar have been changed to icons and moved right.

As discussed in the google groups this is part of the continuing
iterations to try to get some useful style.

It's true that the icons are not immediately understandable,
but they are memorable and learnable (via just trying them and by
tooltips). Learnability rather than totally obvious is sort of at
the core of good tool design (not say this is good, though).