Put a home icon in the upper left
fixes #107 in one way, but not others

The home icon goes to the user's main space if they are logged in,
and to / if this is GUEST.

This is distinct from going to the /dash and may have the potential
of elevating the home space to a central place. We'll see.

Also adjust the css for signing in to use fa-fw and fa-li