Adjust the links for the various ideas of "home"
Fixes #116

This is probably not perfect but at least an improvement. As
@jsavage rightly noticed, there are at least 3 different concepts of
"home" presented on an individual tank page:

* the frontpage of the tank
* the frontpage of "Tank", the entire system
* the tank named after the current user

To address these differences I've made the title text on the
various links more explicit. It's a bit cumbersome and/or wordy now
but at least attempts to explain the difference somewhat.